+++ itest/e2e_test.go
@@ -4,17 +4,19 @@
 package e2etest
 
 import (
-	sdkmath "cosmossdk.io/math"
-	"github.com/babylonlabs-io/babylon/testutil/datagen"
-	"github.com/babylonlabs-io/finality-provider/clientcontroller"
-	"github.com/babylonlabs-io/finality-provider/finality-provider/cmd/fpd/daemon"
-	"github.com/babylonlabs-io/finality-provider/types"
-	"github.com/btcsuite/btcd/btcec/v2"
-	"github.com/stretchr/testify/require"
-	"go.uber.org/zap"
 	"math/rand"
 	"testing"
 	"time"
+
+	sdkmath "cosmossdk.io/math"
+	"github.com/babylonlabs-io/babylon/testutil/datagen"
+	"github.com/btcsuite/btcd/btcec/v2"
+	"github.com/stretchr/testify/require"
+	"go.uber.org/zap"
+
+	"github.com/babylonlabs-io/finality-provider/clientcontroller"
+	"github.com/babylonlabs-io/finality-provider/finality-provider/cmd/fpd/daemon"
+	"github.com/babylonlabs-io/finality-provider/types"
 )
 
 var (
@@ -84,6 +86,13 @@ func TestDoubleSigning(t *testing.T) {
 
 	finalizedBlocks := tm.WaitForNFinalizedBlocks(t, 1)
 
+	// test duplicate vote which should be ignored
+	res, extractedKey, err := fpIns.TestSubmitFinalitySignatureAndExtractPrivKey(finalizedBlocks[0])
+	require.NoError(t, err)
+	require.Nil(t, extractedKey)
+	require.Empty(t, res)
+	t.Logf("duplicate vote for %d is sent", finalizedBlocks[0].Height)
+
 	// attack: manually submit a finality vote over a conflicting block
 	// to trigger the extraction of finality-provider's private key
 	r := rand.New(rand.NewSource(time.Now().UnixNano()))
@@ -91,19 +100,13 @@ func TestDoubleSigning(t *testing.T) {
 		Height: finalizedBlocks[0].Height,
 		Hash:   datagen.GenRandomByteArray(r, 32),
 	}
-	_, extractedKey, err := fpIns.TestSubmitFinalitySignatureAndExtractPrivKey(b)
+	_, extractedKey, err = fpIns.TestSubmitFinalitySignatureAndExtractPrivKey(b)
 	require.NoError(t, err)
 	require.NotNil(t, extractedKey)
 	localKey := tm.GetFpPrivKey(t, fpIns.GetBtcPkBIP340().MustMarshal())
 	require.True(t, localKey.Key.Equals(&extractedKey.Key) || localKey.Key.Negate().Equals(&extractedKey.Key))
 
 	t.Logf("the equivocation attack is successful")
-
-	tm.WaitForFpShutDown(t)
-
-	// try to start the finality providers and the slashed one should expect err
-	err = tm.Fpa.StartHandlingFinalityProvider(fpIns.GetBtcPkBIP340(), "")
-	require.Error(t, err)
 }
 
 // TestFastSync tests the fast sync process where the finality-provider is terminated and restarted with fast sync
@@ -186,7 +189,7 @@ func TestFinalityProviderEditCmd(t *testing.T) {
 
 	args := []string{
 		fpIns.GetBtcPkHex(),
-		"--" + fpdDaemonAddressFlag, tm.FpConfig.RpcListener,
+		"--" + fpdDaemonAddressFlag, tm.FpConfig.RPCListener,
 		"--" + monikerFlag, moniker,
 		"--" + websiteFlag, website,
 		"--" + securityContactFlag, securityContact,
@@ -217,7 +220,7 @@ func TestFinalityProviderEditCmd(t *testing.T) {
 	moniker = "test2-moniker"
 	args = []string{
 		fpIns.GetBtcPkHex(),
-		"--" + fpdDaemonAddressFlag, tm.FpConfig.RpcListener,
+		"--" + fpdDaemonAddressFlag, tm.FpConfig.RPCListener,
 		"--" + monikerFlag, moniker,
 	}